third_party/blink/web_tests/virtual/mathml-disabled/event-handler-attributes.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
  ["click", "dblclick", "mouseover"].forEach(eventName => {
    const attributeName = `on${eventName}`;
    test(t => {
      const el = document.createElementNS(
        "http://www.w3.org/1998/Math/MathML",
        "math"
      );
      el.setAttribute(attributeName, `window.${eventName}Happened = true;`);
      assert_equals(el[attributeName], null);
    }, `The ${attributeName} property on the <math> element must be null when MathML is disabled.`);
    test(t => {
      const el = document.createElementNS(
        "http://www.w3.org/1998/Math/MathML",
        "math"
      );
      let target = undefined;
      el[attributeName] = (e) => { target = e.currentTarget; }
      el.dispatchEvent(new Event(name));
      assert_equals(target, undefined);
    }, `${eventName} should not be handled by the <math> element when MathML is disabled.`);
  });
</script>
</body>
</html>
